+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  * Roseplot weighting fixed
  * strike method added to planara features
  * from_declination added to defgrad to create rotation correcting magnetic declination
+ * defgrad from_euler, from_quat, euler and quat methods added
 
 ### 1.4.0 (Feb 20 2026)
  * attributes storage implemented (passed as kwargs to features)